LPD-93937 Extract the process name resolution into BackgroundTaskUtil
The export and import process resources each carried an identical private method to resolve a friendly process name, falling back from the configuration name to the portlet display name. Move it to the shared util, which now holds more than the progress percentage and is renamed from BackgroundTaskProgressUtil accordingly.
